Moved all classes related to CooldownFuture to their own package
