+++ src/main/java/dte/cooldownsystem/tasks/WhenOverHandler.java
@@ -10,8 +10,8 @@ import java.util.UUID;
 import org.bukkit.scheduler.BukkitRunnable;
 
 import dte.cooldownsystem.cooldown.Cooldown;
-import dte.cooldownsystem.cooldown.future.CooldownFuture;
 import dte.cooldownsystem.cooldown.listeners.CooldownCreationListener;
+import dte.cooldownsystem.cooldownfuture.CooldownFuture;
 import dte.cooldownsystem.utils.CollectionUtils;
 
 public class WhenOverHandler extends BukkitRunnable implements CooldownCreationListener
